requirements.txt
behave==1.2.5
behave-django==1.0.0
Django==2.0.1
factory-boy==2.10.0
Faker==0.8.12
flake8==3.5.0
mccabe==0.6.1
parse==1.8.2
parse-type==0.4.2
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.2
pytz==2017.3
selenium==3.8.1
six==1.11.0
text-unidecode==1.2
